Add option to use "population-aware" allele frequency filter
There are now two options when filtering by allele frequency.
The "naive" option will filter alleles below a certain threshold
per marker per sample (default). The new "population-allele"
option will retain alleles that are seen in other samples that
passes the threshold, even if the frequency of the same allele
in that particular sample did not reach the threshold.